feat: multicast ingest, source health checks, flussonic service, deploy scripts

+++ app/Services/CacheService/CacheManager.php
@@ -18,7 +18,7 @@ class CacheManager implements CacheManagerInterface
     {
         $this->redisCache = $redisCache;
         $this->defaultDriver = config('cache.default', 'redis');
-        $this->defaultTtl = config('cache.ttl', 3600);
+        $this->defaultTtl = (int) config('cache.ttl', 3600);
     }
 
     public function get(string $key, mixed $default = null): mixed
